// src/openprovider/types.ts
export const RegisterDomainArgs = z.object({
  domain: z.object({ name: z.string().min(1), extension: z.string().min(1) }),
  period: z.coerce.number().int().min(1).max(10),
  owner_handle: z.string().min(1),
  admin_handle: z.string().optional(),
  tech_handle: z.string().optional(),
  billing_handle: z.string().optional(),
  name_servers: z
    .array(
      z.object({
        name: z.string().min(1),
        ip: z.string().optional(),
        ip6: z.string().optional(),
      }),
    )
    .optional(),
  ns_group: z.string().optional(),
  is_private_whois_enabled: z.boolean().optional(),
  is_dnssec_enabled: z.boolean().optional(),
  autorenew: z.enum(['on', 'off', 'default']).optional(),
});
export type RegisterDomainArgs = z.infer<typeof RegisterDomainArgs>;